pyproject.toml
[project]
name = "osu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "dearpygui>=2.1.0",
    "dxcam>=0.0.5",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.3",
    "maturin>=1.9.0",
    "mouse>=0.7.1",
    "mutagen>=1.47.0",
    "numpy>=2.3.1",
    "pandas>=2.3.0",
    "pillow>=11.3.0",
    "polars>=1.31.0",
    "pyautogui>=0.9.54",
    "pygame>=2.6.1",
    "pymem>=1.14.0",
    "python-dotenv>=1.1.1",
    "pywin32>=310",
    "scikit-learn>=1.7.0",
    "tqdm>=4.67.1",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
]

[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu118", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu118", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]